+++ src/main/java/com/ming/dao/RolePermissionMapper.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface RolePermissionMapper extends Mapper<RolePermission> {
 
     @Select("select permission_id from role_permission where role_id = #{roleId}")
-    List<Integer> queryPermissionIdsByRoleId(Integer roleId);
+    List<Long> queryPermissionIdsByRoleId(Long roleId);
 }
